Add notice and test regarding ignored time zone
The function 'get.date.from.string' does currently not handle time
zones, that is, it expects every input to be in UTC, even if there is
another time-zone suffix in the input.

To make the user aware of that, add a notice to this function and add an
additional test which makes sure that the time-zone suffix is, indeed,
ignored.

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>