Introduce the vertex attribute 'kind'
Now, the different kindes of artifacts (Mail, Issue, File, Function,
Feature, Featureexpression, Author) can be plotted with different
colors. For this functionality, a new vertex attribute 'kind' is
necessary, which includes the information of the attribute
'artifact.type' or 'Author' (for author vertices).

The new vertex attribute is respected as vertex color during plotting.

Signed-off-by: Barbara Eckl <ecklbarb@fim.uni-passau.de>
Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>