Change the 'simplify.network' function to handle multi-edge networks
We would respect the edge attribute 'relation' while simplifying a
network, therefore, the new algorithm for simplification is now the
following:
First, the networks are split depending on the relation type. Then, they
are simplified and, in a last step, the networks are merged to one
simplified network which have parallel edges if these have different
relation types.

Also, this works on networks without the 'relation' attribute!

Signed-off-by: Barbara Eckl <ecklbarb@fim.uni-passau.de>
Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>