Enable contraction of imperfect ranges in the end
When constructing ranges, in some situations, we might need to avoid that
the last range is not a perfect range any more, i.e., it is smaller than the
specified time period. Therefore, add a parameter 'imperfect.range.ratio'
which quantifies the allowed imperfectness (e.g., at least 70% of the
specified time period are needed to keep an imperfect range, otherwise it
will be combined with the previous range.)

This fixes #104.

In addition to that, add a parameter 'include.end.date' to specify whether to
add 1 second to the end (for including the end date as the end of ranges is
exclusive) or not. (The default should be to add 1 second to the end, as
hitherto. However, we came across situations where we explicitlydon't want
that behavior, which is the reason for adding the additional parameter.)

Also add some additional tests for the newly introduced behavior.

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>