Adapt 'imperfect.range.ratio' in function 'construct.overlapping.ranges'
To improve the current implementation of the function
'construct.overlapping.ranges' regarding the parameter
'imperfect.range.ratio', we now use one single compact if-block for
correcting the ranges *after* their construction.

The test suite is adapted to catch the corner case where we obtain a
single range which is already shorter than the chosen time period –
while passing also 'imperfect.time.ratio' in a manner that would cause
the adaptation of the last range.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>
Reviewed-by: Thomas Bock <bockthom@fim.uni-passau.de>