Add equals method to the RangeData class
To do so, fix the check for the object type in the equals method of the
ProjectData class to also allow RangeData objects. Call this method from
the equals method in RangeData and then compare the remaining two
data sources (`range` and `revision.callgraph`).
Also add a test for this and fix the faulty documentation of
`get.data.cut.to.same.date()`
This fixes #116

Signed-off-by: Christian Hechtl <hechtl@fim.uni-passau.de>