Update templates for issues und pull requests
For clearer guidelines for the user/developer regarding issues and pull
requests, the corresponding templates are updated. We now distinguish
bug reports and feature requests for issues and represent a checklist
with prerequisites for pull requests.

The template files are now located in the folder '.github/' in the
project's root folder.

The contribution guide is updated accordingly.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>