Add part on functionality to README.md
For better comprehensibility, we add a section on the library's
functionality to the README file. This includes a short list of data
sources (grouped by type), very basic information on network
construction (igraph, network attributes, types of networks), and the
list of available files/modules.

The paragraph on  the various types of networks is not formulated yet,
so this is a work-in-progress patch.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>