Enhance README.md on mandatory attributes
For a better overview and to give more information to the interested
reader/user, the README section on mandatory vertex and edge attributes
is further extended. For each mandatory attribute, we now give
information on the attribute's name, the reasoning and information
stored in the attribute, and, if applicable, possible values.

This fixes #121.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>