Adjust legend orientation and placement in plots
As the legends for the network plots can get really long, a complete
line-oriented layout for them is impractical, esepcially when there are
multiple relations present in the network to be plotted. To mitigate the
problem, the legend is now column-oriented by default. This means,
individually, the multiple sublegends/guides are presented in a column-
oriented way, while the sublegends are positioned next to each other in
a horizontal manner.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>