Refactor 'add.edges.for.bipartite.relation' for better readability
Until now, when adding bipartite edges, we rely on the attribute
"relation" being set on the various bipartite-relation data (as given by
the function 'get.bipartite.relations') rather than the names from the
outer list (set by the same function). Now, we iterate over the names of
'bipartite.relations' in the function 'add.edges.for.bipartite.relation'
instead to enhance readibility and comprehensibility and make use of the
names of the outer list.

Additionally, the documention for the function 'get.bipartite.relations'
is adapted to reflect the names on the resulting list.

This fixes the first part of #118.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>