Remove function 'combine.networks'
After merging PR #115, we can remove the function 'combine.networks' and
rely completely on the functions 'igraph::disjoint_union' and
'add.edges.for.bipartite.relation'. This improves readability and
minimizes the interface of the file 'util-networks.R'.

The function 'igraph::disjoint_union' has been chosen instead of
'merge.networks' because the runtime of the chosen function is most
likely lower in this specific case as the other function is far more
generic.

This fixes #118.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>