Change the `split.networks.time.based` method
This method used to take the time range of the network that contains the
earliest timestamp. This lead to the fact that networks that had later
last timestamps were not splitted entirely but only until the last
timestamp
of the first network.
This is fixed with this commit and the bins are now created from the
first
timestamp to the last timestamp of all networks.

Signed-off-by: Christian Hechtl <hechtl@fim.uni-passau.de>