Remove handling of missing committer data
From now on, we do not support data not containing committer data
anymore. For that reason, we can remove the if clause which adds
three NA columns if the committer data was missing.

This eases aggregating committer data for the future. However, to read
in commit data, the input data must contain all the 16 required columns.

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>