Filter duplicated lines in the commit input-data
When reading the commit data, remove duplicated lines, even if they have
different commit ids but the same commit hash.

In addition, if there are additional lines which are identical except
for the "commit.id" and the "artifact.diff.size", treat them as
duplicated. In that case, keep just one of the commit ids (the lower
one) and calculate the mean of the different "artifact.diff.size"
values. So, in the end, we only get one line out of those duplicated
ones.

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>