Make test data (commits.list) more consistent and update tests
In some test data, the committer data (i.e., committer date, committer
name, and committer e-mail) is still missing. As those data are required
now and are not allowed to be 'NA' any more, all the missing data is
added to the test data.

In addition, there were some inconsistencies between the different
commits.list files in the test data and also within a commits.list file.
This commit fixes all the inconistencies in the commit.list files and
updates all the respective tests. As a result of that, some of the test
cases cover even more possible situations than before.

Important notice: Please keep the test data for commits consistent, that
is, apply changes to the proximity test data also to the feature test
data, and not only for global data but also for range-level data.

This fixes #131.

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>