Improve instructions on vertex attributes in README file
As pointed out in PR #135, the instructions on adding vertex attributes
to constructed networks get more and more confusing due to the
increasing number of helper functions in the file
`util-networks-covariates.R`. To be more precise, the functions to add
vertex attributes are now described with the pattern
"add.vertex.attribute.*" in the README file.

Additionally, some typos are fixed.

Props to @bockthom for pointing this out.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>