Update example plot in README
As the bipartite edges in the exemplary plot has been directed from the
artifacts to the authors (which is wrong!), the plot is updated with all
edges reversed.

Additionally, the code to generate the plot and dump it to disk is added
to the showcase file.

Thanks to @ecklbarb and @bockthom for spotting this.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>