Distinguish directedness of author networks and edge-construction algorithm
From now on, the 'author.directed' parameter in the network
configuration only affects the directedness of an author network, and
the newly introduced parameter 'respect.temporal.order' is responsible
for choosing the edge-construction algorithm.

Now, it is possible to build author networks which are directed but
don't respect the temporal order. That is, for two subsequent events,
there will be two edges between the corresponding nodes, one in each
direction.

In addition, it is also possible to build author networks which are
undirected but respect the temporal order. That is, only one edge
(annotated with the temporally later date) will be present between the
corresponding nodes of two subsequent events.

So, let there be an event of author A to common artifact at time 1 and
an event of B to the same artifact at time 2. Then, we are able to
construct the following networks (only the list of edges is presented
here):

author.directed=TRUE, respect.temporal.order=TRUE:   A<--2--B
author.directed=TRUE, respect.temporal.order=FALSE:  A--1-->B, A<--2--B
author.directed=FALSE, respect.temporal.order=TRUE:  A--2-- B
author.directed=FALSE, respect.temporal.order=FALSE: A--1--B, A--2--B

For more information, please see #6.

To keep the former behavior, the default value of the newly introduced
network configuration parameter 'respect.temporal.order' is NA. If the
default value this parameter is used, the edge-construction function
uses the 'directed' parameter for 'respect.temporal.order'.

This fixes #6.

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>