Deal with empty data sources in 'get.first.activity.data'
When the data for a data source are empty (i.e., an empty data.frame),
we need to deal with the corresponding empty list 'minimums.per.person'
in the function 'get.first.activity.data'; otherwise, the next lines of
code will throw an error.

The easy fix is to check whether the data are empty, add NA values to
the intermediate data structure and, then, skip the next lines of code.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>