Replace 'get.first.activity.data' with faster implementation
When planning the functionality for cmputing the vertex attribute
'first.activity', the straight-forward idea was to "merge lists" and
then simply run 'min' for the global first activity if needed. When
implementing the functionality in commit
334c2585a03548baf840a38db079f69cb48d4e70, the "simple" merge has proven
to be not that easy to perform.

With some effort, in this patch, we replace the current implementation
with the originally planned version.

When performing a microbenchmark on BusyBox sample data (five two-year
ranges of data and networks), comparing both versions of the code (see
column 'expr' below), the new version proves to be faster (also due to
better parallelization):
expr      min       lq     mean   median       uq      max neval
   old 83.07881 84.16226 85.79368 85.44123 87.45646 89.70994    20
   new 75.85983 78.09042 78.76604 78.71732 79.44771 83.42560    20
Additionally, the code is more concise and we do not need additional
helper functions.

Add namespace imports to module 'networks-covariates'.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>