Adapt 'ProjectData$get.artifacts' to work with all data sources
When calling the function 'ProjectData$get.artifacts()', we currently
retrieve the list of source-code artifacts only. With this patch, we can
give the desired data source as a parameter to the function.

To implement this generic functionality, we need to add two global map
constants: 'DATASOURCE.TO.ARTIFACT.FUNCTION' and
DATASOURCE.TO.ARTIFACT.COLUMN. The first maps the data sources to the
specific function that is needed to get the correct (filtered) artifact
data. The latter maps the data sources to the specific column in the
data where the artifacts are stored.

This puts work on issue #97.

Additionally, adjust some TODO items.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>