Determine list of artifacts more reasonably in ProjectData
As a follow-up to 0d184b82dac1a6d7edb3a3339e8352343cbe020f, the
data-mapping methods of ProjectData are adapted as follows.

The following methods are incorporated into 'get.author2artifact' and
'get.artifact2author', respectively, where the data source needs to be
passed as a parameter:
- get.thread2author,
- get.author2thread,
- get.issue2author, and
- get.author2issue.

The following method is incorporated into 'get.artifact2artifact' which
retrieves data for the bipartite relations (the data source needs to be
passed as a parameter):
- get.commit2artifact.

The following methods are removed:
- get.author2commit,
- get.author2mail,
- get.author2file, and
- get.commit2file.

Affected methods and functions as well as the showcase file are adapted
appropriately. The test suite is fixed due to a re-order of vertices in
one test case.

This fixes #97.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>