Split data into time-based equally-sized windows
Currently, we have the possibility to split data in a time-based manner
by specifying a time period or specific bins. However, we do not have
the possibility to specify the number of windows. With this commit, we
add this functionality.

To implement this functionality, the functions 'split.data.time.based'
and 'split.get.bins.time.based' both get a new parameter
'number.windows' and the function 'generate.date.sequence' a parameter
'length.out'.

Additionally, adjust function documentation appropriately.

This fixes #49.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>