Fix function 'split.networks.time.based'
The 'sapply' in the function 'split.networks.time.based' does not work if
the provided list of networks only contains one element: In that case,it
returns a matrix containing one column instead of returning a list.
However, this leads to problems when passing this matrix to the function
'get.date.from.unix.timestamp'.
When using 'lapply' instead, we get a list all the time, which fixes
the problem.

In addition, the corresponding test in the test suite is enhanced by an
additional statement which just runs the function with a list containing
one element (to keep track of that for future changes).

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>