Improve function 'save.and.load'
The function 'save.and.load' should also be able to work without
assignment of its return value. This way, the real potential of the
parameter 'variable' is unleashed. Until now, the parameter was needed
to access the given variable from the environment stored in a dumped
file.

With the solution implemented in this patch, the stored variable is
assigned to the calling environment under the given name (independent
from the data's existence). This way, the assignment of the return value
is not necessary anymore. However, for compatibility, it is still
possible to assign the return value to another variable name when
calling the function.

Additionally, the documentation of the function is improved by giving
more details on the internals and a (hopefully) clearer wording.

Thanks to @hechtlC for pointing this out, although unintentionally.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>