Split networks into time-based equally-sized windows
Currently, we have the possibility to split data in a time-based manner
by specifying a time period, specific bins, or the number of resulting
time windows. However, we do not have the latter possibility for network
splitting. With this commit, we add this functionality.

To implement this functionality, the functions
'split.network.time.based' and 'split.networks.time.based' both get a
new parameter to handle this functionality. Additionally, streamline
other 'split.*' functions for uniformity regarding the 'number.windows'
and 'sliding.window' parameters.

Finally, adjust function documentation appropriately.

This is a follow-up for commit 40974bada77b9184b025e47ebd7f0c4fb98e1475.
Hopefully, this really fixes #49.
Thanks to @bockthom for pointing this out in his review on PR #140.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>