Handle ARPACK errors in eigen_centrality calculation
For some reason, when using the parameter 'directed = TRUE', the function
'igraph::eigen_centrality' can run into ARPACK errors (in some rare cases).

For example, for a small subnet of an author network of BusyBox having
7 nodes and 12 edges, we get the following error:

Error in base::.Call(.NAME, ...) :
  At arpack.c:1174 : ARPACK error, Maximum number of iterations reached
In addition: Warning message:
In .Call("R_igraph_eigen_adjacency", graph, algorithm, which, options, :
  At arpack.c:776 :ARPACK solver failed to converge (1001 iterations,
0/1 eigenvectors converged)

Those errors were already discussed in the igraph project on GitHub, e.g.:
https://github.com/igraph/igraph/issues/512
According to the conversation there, to avoid those errors, the number of
iterations could be increased from its default value 1000 to a higher number
of iterations. Also the ncv value could be increased. However, in the above
example, this still runs into to the described error.

Hence, we need a more drastic action: Increase the accuracy tolerance
from 0 to 0.1. These leads to less precise results but omits the error
(at least in the example stated above).

So, in general, we still compute 'eigen_centrality' with its default
ARPACK options. If this leads to an error, compute the 'eigen_centrality'
again by using the adjusted ARPACK options (maxiter = 5000 and tol=0.1).

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>