Fix problem with fractional time periods in 'generate.date.sequence'
When giving a parameter 'length.out' to the function
'generate.date.sequence', the resulting time period may get a fractional
number. This may result in additional generated dates, i.e., the length
of the date sequence is not 'length.out' anymore – with the difference
between the last two dates being likely only one second.

To fix this, the calculated time period is rounded up. This will
decrease the length of the last time period between dates, but, at most,
only by 'length.out' seconds.

To introduce a test case for this, the function
'split.get.bins.time.base' is fixed to set the levels on the bin factor
in the correct order.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>