Fix sliding windows in splitting functionality
Due to an error introduced in 94cc87bd1789b211aff13b7c26235ccee77c2b7b,
the sliding-window parameter has not been handled correctly when
splitting networks and data. To fix the problem, the code to disable
sliding windows when splitting bins are set, is moved to the correct
position.

Additionally, the code in the function 'split.network.time.based' is
adapted to conform with the other functions regarding the if-condition
on specified bins.

Props to @bockthom for spotting this.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>
Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>