Allow merge of empty networks (edges and/or vertices)
The functions 'merge.networks' and 'merge.network.data' should work in a
generic way, such that they work also on general-purpose networks not
coming from the network library. To enable this, the case that empty
networks (in respect to number of edges and/or vertices) is now handled
properly.

In detail, two special cases need to be handled:
- empty edge lists created from edgeless networks need to be filtered
out and
- if all edge lists end up empty, an appropriate empty edge list needs
to be created (instead of NULL).

Passing over empty vertex data (which results in 'NULL' in function
'merge.network.data') does not affect the correct handling of data right
away.

Appropriate test cases are added in the file 'tests/test-networks.R'.

This fixes #142.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>