Add possibility of multiple activity types to active ranges
The function add.vertex.attribute.active.ranges is adapted. It is possible now
to call it with a vector of multiple activity types. The lists containing active
ranges per type and author as before are combined to a list containing an element
for every given activity type and added as vertex attribute.

Signed-off-by: Klara Schlueter <schluete@fim.uni-passau.de>