Handle incorrect keys and values in 'get.key.to.value.from.df'
When splitting and subsetting data with the function
'get.key.to.value.from.df', it may happen that the given values for the
parameters 'key' and 'value' are not valid column names of the
'base.data'. Whenever a invalid specification is given, an error message
(including stack trace) is emitted and the script is stopped.

A handling in this function automatically enables the analogous handling
in the method 'ProjectData$group.data.by.column'.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>
Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>