Change shape of 'Vertices' in the legend of plots
The shape of the legend for 'Vertices' was the same as the shape of the
'Author' vertex type. This is a little bit confusing when we have two
different vertex types and vertices in the network:
Then the shape of 'Vertices' in the legend is the same as for the
vertex type 'Author' even if the vertex type is 'Artifact'.

To avoid confusion, use another shape in the legend for 'Vertices'
which is (usually) not used for vertex types. Then it is more clear that
not the shape of 'Vertices' in the legend does matter, but the color.

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>