Update exemplary multi network in README.md
Update the exemplary multi network, which is displayed in the README.md, to
contain the shape changes in the legend
(see f4fb4807cfd87d9d552a9ede92ea65ae4a386a04).

In addition, add width and height parameters to the ggsave statement in the
showcase.R file which generates this examplary multi network.

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>