Adjust read.commits to return a valid data.frame instead of an empty one
In the case of a valid commits.list file with at least one commit line the
read.commits function returns a data.frame with 16 columns containing all the
commits read from the file. If the commits.list file is empty, however, it
previously returned an empty data.frame with no columns. This has now been
adjusted to return an empty data.frame with all the columns (16 columns), which
should save a lot of additional if-else case distinctions later on because now
the shape of the returned data.frame by the read.commits function is always the
same.

Signed-off-by: Jakob Kronawitter <kronawij@fim.uni-passau.de>