Restructure get.commits and get.commits.filtered(.empty) methods
This major commit merges the two old methods get.commits.filtered and
get.commits.filtered.empty of the ProjectData class into one new method again
called get.commits.filtered. Similiarly, the filter.commits.empty and
filter.commits methods were merged into one new filter.commits method which now
takes filter.untracked.files and artifact.filter.base as paramaters which then
control how the filtering is performed.

The filter.untracked.files parameter was added to the ProjectConf which now
controls - just like the artifact.filter.base parameter - which commits should
be filtered out when calling the get.commits.filtered method.

If you want to call the get.commits.filtered with other paramaters (not the ones
that are configured in the ProjectConf) then one can call the
get.commits.filtered.uncached version of this method. As the name implies, this
method is not taking advantage of caching and should thus not be used too often.

In the course of revamping these methods it only took a minor effort to rename
the empty artifact to a more speaking identifier, namely, "untracked files".
Thus, this renaming was also performed in this commit.

Signed-off-by: Jakob Kronawitter <kronawij@fim.uni-passau.de>