Adapt test cases to new changes and improve empty dataframe creation
The test cases were adapted to two of the new changes in the network library.
The first one is the fact that the get.commits method now removes either
'Feature' or 'FeatureExpression' commits. The second one was the change that
there are no dummy data.frames anymore (with zero columns and rows). Instead
there are empty data.frames when there no data exists (with columns but zero
rows). One mistake was made during creation of these. The empty data.frames
that are created did not contain any data type informtion (all columns defaulted
to the 'logical' data.type). If this is not wanted there now exists a new helper
method which also takes care of data types.

Signed-off-by: Jakob Kronawitter <kronawij@fim.uni-passau.de>