Change edge generation behaviour for base and untracked files artifact
Previously, when an author network was created and the untracked files artifact
and the base artifact were included, edges have been created among the untracked
files artifact and among the base artifact. This was now changed so that there
are no edges created among untracked files at any time. For the base artifact it
can be configured via the new base.artifact.edges parameter in the NetworkConf.

Signed-off-by: Jakob Kronawitter <kronawij@fim.uni-passau.de>