Update default behavior of 'Conf' objects
In recent scenarios and in perspective of up-coming changes, the default
behavior of 'Conf' objects upon initialization and update:

1) The default values are *not* automatically checked against the
allowed values anymore. This is mainly disabled to avoid confusion of
users. The constructor of the class 'Conf' is adapted accordingly.
2) When updating a configuration value, the program execution is now
stopped (using 'stop') on failure. Previously, the respective update has
been ignored while issuing a warning. This change helps preventing
confusion and analysis errors early in an analysis script. Accordingly,
the parameter 'stop.on.error' to all update methods is removed.

Furthermore, the code is streamlined, such that the super-constructor is
called from both subclasses 'NetworkConf' and 'ProjectConf'. Some log
statements are added/adjusted, too.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>