Leave artifact column empty if artifact == file or artifact == funtion
For commits to untracked files the artifact column has previously been the
copied file column (for example when looking at the commit data returned by
'get.commits'). However this should only be the case when considering file level
analysis (e.g. 'artifact == file' in the 'ProjectConf').
This commit changes this to the correct behaviour. So for 'artifact == function'
and 'artifact == feature' the artifact column now only contains the empty string
for untracked files. To avoid hardcoding this empty string in every affected
place a global constant called 'UNTRACKED.FILE.EMPTY.ARTIFACT' was added.

Signed-off-by: Jakob Kronawitter <kronawij@fim.uni-passau.de>