Fix hierarchy calculations and classifications
In the function 'metrics.hierarchy', the values for all vertices that
are NaN or 0 get removed. On the one hand, this violates the function
documentation, on the other hand, this prevents the correct calculation
of the hierarchy-based developer classification.
To fix this issue, in the function 'metrics.hierarchy', the values are
not filtered anymore. Additionally, the documentation of the function
'metrics.clustering.coeff' is adapted and extended in the same way as
the one of 'metrics.hierarchy'.

Regarding the hierarchy-based author classification, the implementation
of the function 'get.author.class.network.hierarchy' is adapted
accordingly by replacing hierarchy values (i.e., vertex degree divided
by clustering coefficient) of 'Inf' with '0'.
There are no complications here because there are no authors with a
vertex degree == 0 and a clustering coefficient (CC) > 0 (i.e., the
hierarchy value would really be 0 for those). Authors with a CC == NaN
(degree < 2) will stay with their hierarchy value of NaN, accordingly.

Furthermore, a related bug in function 'get.author.class' is fixed:
'drop = FALSE' cannot be given in a data.frame assignment on the
left-hand side.

Thanks to @bockthom for the related discussion on this patch.

This fixes #151.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>
Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>