Add additional utility functions for easier empty dataframe creation
The following four functions are introduced to make empty dataframe creation
even easier:
- create.empty.authors.list
- create.empty.commits.list
- create.empty.issues.list
- create.empty.mails.list

As the name implies, each function is creating and returning an dataframe with
zero rows but the correct column names and datatypes (like a dataframe of the
corresponding type would have if it was filled with authors, commits, issues or
mails, respectively).

In addition, all the column names and column datatypes are made publicly
available by defining them as constants.

Signed-off-by: Jakob Kronawitter <kronawij@fim.uni-passau.de>