Adjust comments for the column names of commonly used dataframes
Previously, next to the declaration of column names for commonly used dataframes
(dataframes containing commits, mails, etc.), there was an SQL statement
describing how the data was retrieved before it is read by the network library
(in the tool 'codeface-extraction').
These SQL statements are now removed and instead, it is referred to the tool
'codeface-extraction' itself on GitHub.

Signed-off-by: Jakob Kronawitter <kronawij@fim.uni-passau.de>