Rework 'tests.R' and fix logger initalization for tests
As newer versions of the package 'logging' (>= 0.8) are sensitive to the
attachment of the base package 'methods', the latter package is now
loaded explicitly during initialization and for the test suite. The
underlying reason is that the package 'methods' is not loaded when
running Rscript (see '?Rscript' for details).

For debugging purposes, the output of the test suite is augmented with
session information and log statements for the individual stages that
are executed before running the test suite.

This patch fixes the original part of issue #152.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>