Fix merging of networks without vertices
The function 'merge.networks' is not able to handle networks that have
not vertices and also (!) no vertex attributes: The final call to
'igraph::graph.data.frame' results in an error.
To fix this, the special case is identified and the list of vertices is
adjusted accordingly. A corresponding test is added to the test suite.

Additionally, the default case of merging only one network is handled by
returning the only network right away.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>