Remove the empty artifact from all types of networks
Augmenting the changes from PR #149, the empty artifact, which results
from untracked files, is not a part of any network type anymore.

This is change is based on commit d9f527c5602a3f463e5ccb0d395abe1d6a837ea3
and fixes #153.

This is a preliminary implementation with regard to issue #154.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>