Add parameter 'restrict.classification.to.authors' and clean up code
This commit adds the new parameter 'restrict.classification.to.authors' to the
following functions:
- 'get.author.class.network.degree'
- 'get.author.class.network.eigen'
- 'get.author.class.network.hierarchy'
- 'get.author.class.commit.count'
- 'get.author.class.loc.count'
- 'get.author.class.by.type'
- 'get.author.class.overview'

If specified (the parameter can be omitted, i.e. is optional), a list of author
names has to be passed. The classification will then be performed only on the
authors that exist in the respective datasource (the network for network-based
classifications or the project data for count-based classifications). Authors
that do exist in the datasource but are not specified in the parameter, will be
ignored. Authors that are specified in the parameter but do not exist in the
datasource are appended to the finished classification results. They get
appended as peripheral author and with a centrality value (for example, this
could be a commit-count or hierarchy value) of 'NA'.

The need of this parameter was discussed in issue #148.

In addition to introducing this parameter, lots of redundant code is
eliminated. This is achieved by moving the specific classification funtionality
for each type (e.g. network.degree, commit.count, etc.) into the function
'get.author.class.by.type'. The first five of the seven functions listed above
now forward calls by calling the revised 'get.author.class.by.type'
function internally.

Some minor adjustments are made to dataframe-SQL statements.

Signed-off-by: Jakob Kronawitter <kronawij@fim.uni-passau.de>