Fix reading of mail data
When reading mail data, the setting of the column 'artifact.type' can
lead to errors since the column does not occur in the original data, but
is explicitly added in 'read.mails'. In addition, the constants
'MAILS.LIST.COLUMNS' and 'MAILS.LIST.DATA.TYPES' missed that column.

Moreover, the type of the column 'thread' is rather a character string
than a numeric.

The tests are adapted accordingly.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>
Signed-off-by: Christian Hechtl <hechtl@fim.uni-passau.de>