Fix reading of issue data
When reading issue data, the setting of the column 'artifact.type' can
lead to errors since the column does not occur in the original data, but
is explicitly added in 'read.issue'. Same holds for the column
'event.id' that is generated by a hashing algorithm. In addition, the
constants 'ISSUES.LIST.COLUMNS' and 'ISSUES.LIST.DATA.TYPES' missed both
columns.

Moreover, the columns 'issue.type', 'issue.resolution',
'issue.components', and 'event.info.2' are updated to have the type
'list()' now (the function 'create.empty.data.frame' is adapted
accordingly to handle the type 'list()'). The parsing from JSON for the
column 'event.info.2' is simplified.

Finally, use the function 'create.empty.issues.list' in the method
'ProjectData$set.issues'.

The tests are adapted accordingly.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>