Heavily improve addition of PaStA data
When enabling the 'ProjectConf' parameter 'pasta', the project's PaStA
data is added to commit and mail data when retrieving those. Formerly,
the PaStA items have been added one-by-one, heavily impacting the
performance. In this patch, the addition of PaStA data is
extraordinarily improved.

The changes in this patch in detail:

1) Compute aggregations of the PaStA data (on commit-hash and message-ID
columns independently) to obtain mergeable data.frames for the addition
to the main data sources. The aggregation is performed using the base
function 'aggregate' in the method 'ProjectData$aggregate.pasta.data'.
The results are stored in appropriate fields of the class 'ProjectData'.

2) Adjust the methods 'ProjectData$get.pasta', 'ProjectData$set.commits'
and 'ProjectData$set.mails' accordingly to perform a join with the
aggregated PaStA data (using the function 'merge').

3) Remove the method 'ProjectData$add.pasta.data' as it is obsolete now.

4) Add constants 'PASTA.LIST.COLUMNS' and 'PASTA.LIST.DATA.TYPES' as
well as the function 'create.empty.pasta.list', analogously to the
functions and constants for the other data sources.

5) Apply some minor changes (re-ordering of fields, clearer formatting,
adding/adjusting log statements, etc.).

Overall, the addition process is improved extensively: Before, the
retrieval of commit data for the project 'uboot' (with added PaStA
data!) took about 3:40 minutes. With the changes of this patch, the same
procedure takes about 40 seconds – which is only about 20% of the
previous execution time.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>
Signed-off-by: Christian Hechtl <hechtl@fim.uni-passau.de>